requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
mypy==0.800
mypy-extensions==0.4.3
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
